+++ react/__mocks__/vtex.render-runtime.tsx
@@ -1,0 +1,7 @@
+export const useRuntime = () => {
+  return {
+    culture: {
+      country: 'USA',
+    },
+  }
+}